\name{cap-package}
\alias{cap-package}
\alias{cap}
\docType{package}
\title{ Covariate Assisted Principal (CAP) Regression for Covariance Matrix Outcomes
}
\description{
cap package performs Covariate Assisted Principal (CAP) Regression for covariance matrix outcomes. The method identifies the optimal projection direction which maximizes the log-likelihood function of the log-linear heteroscedastic regression model in the projection space.
}
\author{
Yi Zhao, Johns Hopkins University, <zhaoyi1026@gmail.com> 

Bingkai Wang, Johns Hopkins University, <bwang51@jhmi.edu>

Stewart Mostofsky, Johns Hopkins University, <mostofsky@kennedykrieger.org> 

Brian Caffo, Johns Hopkins University, <bcaffo@gmail.com> 

Xi Luo, Brown University, <xi.rossi.luo@gmail.com> 

Maintainer: Yi Zhao <zhaoyi1026@gmail.com>
}
\keyword{ package }
