\name{Ep}
\alias{Ep}
\title{Ep calculates the expectation of tr(A\%*\%B) based on a permutation test}
\usage{
Ep(A, B)
}
\arguments{
\item{A}{a centralized and symmetric matrix with nrow(A)>1}

\item{B}{a centralized and symmetric matrix with the same dimension as A}
}
\value{
the expectation of tr(A\%*\%B) based on a permutation test
}
\description{
Ep calculates the expectation of tr(A\%*\%B) based on a permutation test
}
