% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugPathway.R
\name{getDrugExposureData}
\alias{getDrugExposureData}
\title{get drug_exposure table for creating sequence data}
\usage{
getDrugExposureData(
  connectionDetails = connectionDetails,
  cdmDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema = cohortDatabaseSchema,
  cohortTable = cohortTable,
  cohortId = cohortId,
  StartDays = StartDays,
  EndDays = EndDays,
  conceptSets = conceptSets
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{cohortId}{cohortId}

\item{StartDays}{start days}

\item{EndDays}{end days}

\item{conceptSets}{concept Sets}
}
\description{
get drug_exposure table for creating sequence data
}
\details{
Extract concept_id from the json file of concept set
}
