% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugPathway.R
\name{totalN}
\alias{totalN}
\title{get drug_exposure table for creating sequence data}
\usage{
totalN(connectionDetails, cohortDatabaseSchema, cohortTable, cohortId)
}
\arguments{
\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{cohortId}{cohortId}
}
\description{
get drug_exposure table for creating sequence data
}
\details{
Extract concept_id from the json file of concept set
}
