\name{Tp}
\alias{Tp}
\title{Tp calculates the third moment of tr(AB) based on a permutation test}
\usage{
Tp(A, B)
}
\arguments{
\item{A}{a centralized and symmetric matrix with nrow(A)>5}

\item{B}{a centralized and symmetric matrix with the same dimension as A}
}
\value{
the third moment of tr(A%*%B) based on a permutation test
}
\description{
Tp calculates the third moment of tr(AB) based on a permutation test
}
