\name{Vp}
\alias{Vp}
\title{Vp calculates the variance of tr(AB) based on a permutation test}
\usage{
Vp(A, B)
}
\arguments{
\item{A}{a centralized and symmetric matrix with nrow(A)>3}

\item{B}{a centralized and symmetric matrix with the same dimension as A}
}
\value{
the variance of tr(AB) based on a permutation test
}
\description{
Vp calculates the variance of tr(AB) based on a permutation test
}
