\name{PVD}
\alias{PVD}
\title{PVD: A fast computational procedure to detect the genetic pleiotropy effect for multiple variants}
\usage{
PVD(G, E)
}
\arguments{
\item{G}{the similarity matrix for genetic data with nrow(G)>5}

\item{E}{the similarity matrix for phenotypic data with nrow(E)=nrow(G)}
}
\value{
the p value for the fast method
}
\description{
PVD: A fast computational procedure to detect the genetic pleiotropy effect for multiple variants
}
